+++ profiler/interface.py
@@ -13,39 +13,49 @@ wlan interface data class
 """
 
 import logging
-# standard library imports
 import os
-import re
 import subprocess
-import sys
+# standard library imports
+from collections import namedtuple
+from copy import copy
+from typing import Dict, List
 
 # app imports
 from .constants import _20MHZ_CHANNEL_LIST
-from .helpers import run_cli_cmd
+from .helpers import flag_last_object, run_cli_cmd
+
 
 class InterfaceError(Exception):
     pass
 
+
 class Interface:
     """WLAN Interface data class"""
 
-    def __init__(self, interface, channel=None, no_interface_prep=False):
+    def __init__(self):
         self.log = logging.getLogger(self.__class__.__name__.lower())
-        self.name = interface.lower()
+        self.name = ""
+        self.frequency = ""
+        self.no_interface_prep = False
+
+    def setup(self):
         self.check_interface(self.name)
-        self.channel = channel
-        self.no_interface_prep = no_interface_prep
+        self.channel = self.get_channel()
         if not self.channel:
-            self.channel = self.get_channel()
-            if not self.channel:
-                self.log.warning("could not determine channel")
+            self.log.warning("could not determine channel")
         self.mac = self.get_mac().lower()
         self.mode = self.get_mode().lower()
+        if self.mode not in ("managed", "monitor"):
+            raise InterfaceError("%s is mode is not managed or monitor" % self.name)
         self.operstate = self.get_operstate().lower()
+        self.phy_id = self.get_phy_id()
+        self.phy = f"phy{self.phy_id}"
+        self.mon = f"mon{self.phy_id}"
         self.driver = self.get_driver()
         self.driver_info = self.get_ethtool_info()
         self.driver_version = self.get_driver_version()
         self.firmware_version = self.get_firmware_version()
+        self.copy = copy(self)
         self.checks()
         self.log_debug()
 
@@ -65,56 +75,156 @@ class Interface:
             self.log.debug("reg domain set to %s", " ".join(regdomain))
             self.log.debug("see 'iw reg get' for details")
 
-    def stage_interface(self) -> None:
-        """Prepare the interface for monitor mode and injection"""
-        if self.mode in ("managed", "monitor"):
-            iwlwifi_scan_commands = [
-                ["ip", "link", "set", f"{self.name}", "down"],
-                ["iw", "dev", f"{self.name}", "set", "managed"],
-                ["ip", "link", "set", f"{self.name}", "up"],
-                ["iw", f"{self.name}", "scan"],
-            ]
-            staging_commands = [
-                ["wpa_cli", "-i", f"{self.name}", "terminate"],
-                ["ip", "link", "set", f"{self.name}", "down"],
-                ["iw", "dev", f"{self.name}", "set", "monitor", "none"],
-                ["ip", "link", "set", f"{self.name}", "up"],
-                ["iw", f"{self.name}", "set", "channel", f"{self.channel}", "HT20"],
-            ]
-            if "iwlwifi" in self.driver:
-                commands = iwlwifi_scan_commands + staging_commands
-            else:
-                commands = staging_commands
-            try:
-                for cmd in commands:
-                    cp = subprocess.run(
-                        cmd, encoding="utf-8", shell=False, capture_output=True
-                    )
-                    if cp.stderr:
-                        if "monitor" in cmd:
-                            cp = subprocess.run(
-                                ["iw", "dev", f"{self.name}", "set", "type", "monitor"],
-                                encoding="utf-8",
-                                shell=False,
-                                capture_output=True,
-                            )
-                        if "wpa_cli" not in cmd:
-                            raise OSError(
-                                f"problem running '{' '.join(cmd)}'\n{cp.stderr}"
-                            )
-                return True
-            except OSError:
-                self.log.exception(
-                    "error setting %s interface config: %s",
-                    self.name,
-                    "\n".join(
-                        [line for line in cp.stderr.split("\n") if line.strip() != ""]
-                    ),
-                    exc_info=None,
-                )
+    def run_command(self, command) -> str:
+        try:
+            cp = subprocess.run(
+                command, encoding="utf-8", shell=False, capture_output=True
+            )
+            return cp.stdout
+        except OSError:
+            raise InterfaceError("problem running %s: %s", command, cp.stderr)
 
-        self.log.error("failed to prepare the interface...")
-        sys.exit(-1)
+    def run_commands(self, commands) -> None:
+        try:
+            for cmd in commands:
+                cp = subprocess.run(
+                    cmd, encoding="utf-8", shell=False, capture_output=True
+                )
+                if cp.stderr:
+                    if "monitor" in cmd:
+                        cp = subprocess.run(
+                            ["iw", "dev", f"{self.name}", "set", "type", "monitor"],
+                            encoding="utf-8",
+                            shell=False,
+                            capture_output=True,
+                        )
+        except OSError:
+            msg = (
+                "error setting %s interface config: %s",
+                self.name,
+                "\n".join(
+                    [line for line in cp.stderr.split("\n") if line.strip() != ""]
+                ),
+            )
+            self.log.exception(msg, exc_info=None)
+            raise InterfaceError(msg)
+
+    def restore_interface(self) -> None:
+        """Delete monitor interface and restore interface"""
+        commands = [
+            ["iw", "dev", f"{self.mon}", "del"],
+            # ["iw", "phy", f"{self.phy}", "interface", "add", f"{self.name}", "type", "managed"],
+            ["ip", "link", "set", f"{self.name}", "down"],
+            ["iw", "dev", f"{self.name}", "set", "type", "managed"],
+            ["ip", "link", "set", f"{self.name}", "up"],
+        ]
+        self.run_commands(commands)
+
+    def scan(self) -> None:
+        """Perform scan in attempt to enable a disabled channel"""
+        iwlwifi_scan_commands = [
+            ["ip", "link", "set", f"{self.name}", "down"],
+            ["iw", "dev", f"{self.name}", "set", "type", "managed"],
+            ["ip", "link", "set", f"{self.name}", "up"],
+            ["iw", f"{self.name}", "scan"],
+        ]
+        self.log.debug("performing scan on %s", self.name)
+        self.run_commands(iwlwifi_scan_commands)
+
+    def get_channels_status(self) -> Dict:
+        iw_version = self.run_command(["iw", "--version"])
+        ip_version = self.run_command(["ip", "-V"])
+        self.log.debug("%s", iw_version.strip())
+        self.log.debug("%s", ip_version.strip())
+        iw_phy_channels = self.run_command(["iw", "phy", f"{self.phy}", "channels"])
+        freq = ""
+        ch = ""
+        no_ir = False
+        band = ""
+        disabled = False
+        first_band = True
+        first_channel_in_band = True
+        bands = {}
+        channels = []
+        channel = namedtuple("channel", ["freq", "ch", "no_ir", "disabled"])
+
+        for line, last_line in flag_last_object(iw_phy_channels.splitlines()):
+            line = line.strip().lower()
+            if first_band:
+                first_band = False
+                if line.startswith("band "):
+                    band = line.split(" ")[1]
+                    continue
+            if first_channel_in_band:
+                first_channel_in_band = False
+                if line.startswith("*"):
+                    freq = line.split()[1]
+                    ch = line.split()[3].replace("[", "").replace("]", "")
+                    continue
+            if line.startswith("*"):
+                channels.append(channel(freq, ch, no_ir, disabled))
+                # reset vars
+                freq = ""
+                ch = ""
+                no_ir = False
+                disabled = False
+                if "disabled" in line:
+                    disabled = True
+                freq = line.split()[1]
+                ch = line.split()[3].replace("[", "").replace("]", "")
+                continue
+            if line.startswith("band "):
+                channels.append(channel(freq, ch, no_ir, disabled))
+                bands[band] = channels
+                # reset channels list
+                channels = []
+                # reset channel flag
+                disabled = False
+                first_channel_in_band = True
+                band = line.split(" ")[1]
+            if line.startswith("no ir"):
+                no_ir = True
+                continue
+            if last_line:
+                channels.append(channel(freq, ch, no_ir, disabled))
+                bands[band] = channels
+        return bands
+
+    def stage_interface(self, freq: str) -> None:
+        """Prepare the interface for monitor mode and injection"""
+        wpa_cli_version = self.run_command(["wpa_cli", "-v"])
+        self.log.debug("%s", wpa_cli_version.splitlines()[0])
+
+        # if channel is disabled, a scan may enable it (iwlwifi like AX210)
+        for _band, channels in self.get_channels_status().items():
+            for ch in channels:
+                if freq == ch.freq:
+                    if ch.disabled:
+                        self.scan()
+                    break
+            else:
+                continue
+            break
+
+        staging_commands = [
+            ["wpa_cli", "-i", f"{self.name}", "terminate"],
+            ["ip", "link", "set", f"{self.name}", "down"],
+            [
+                "iw",
+                f"{self.phy}",
+                "interface",
+                "add",
+                f"{self.mon}",
+                "type",
+                "monitor",
+                "flags",
+                "none",
+            ],
+            ["ip", "link", "set", f"{self.mon}", "up"],
+            ["ip", "link", "set", f"{self.name}", "down"],
+            ["iw", f"{self.mon}", "set", "freq", f"{freq}", "HT20"],
+        ]
+        self.run_commands(staging_commands)
 
     def checks(self) -> None:
         """Perform self checks and warn as neccessary"""
@@ -202,14 +312,21 @@ class Interface:
 
     def get_channel(self):
         """Determine what channel the interface is set to"""
-        iwconfig = run_cli_cmd(["iwconfig"])
-        for line in iwconfig.splitlines():
-            line = line.lower()
-            if self.name in line:
-                if "freq" in line:
-                    _result = re.search(r"frequency:\d+.\d+", line)
-                    _freq = int(float(_result.group().split(":")[1]) * 1000)
-                    return _20MHZ_CHANNEL_LIST[_freq]
+        iw_dev_iface_info = self.run_command(["iw", "dev", f"{self.name}", "info"])
+        for line in iw_dev_iface_info.splitlines():
+            line = line.lower().strip()
+            if "channel" in line:
+                channel = int(line.split(",")[0].split(" ")[1])
+                freq = int(line.split(",")[0].split(" ")[2].replace("(", ""))
+                resp = _20MHZ_CHANNEL_LIST.get(freq, 0)
+                if channel != resp:
+                    self.log.warning(
+                        "iw reported a different channel (%s) than our lookup (%s)", channel,
+                        resp,
+                    )
+                self.log.debug("channel is %s and freq is %s ", resp, freq)
+                return resp
+        return _20MHZ_CHANNEL_LIST.get(self.frequency, 0)
 
     def get_operstate(self) -> str:
         """
@@ -228,6 +345,65 @@ class Interface:
         """
         operstate = run_cli_cmd(["cat", f"/sys/class/net/{self.name}/operstate"])
         return operstate.strip()
+
+    def build_iw_phy_list(self) -> List:
+        iw_devs = run_cli_cmd(["iw", "dev", f"{self.name}", "info"])
+        phy = namedtuple("phy", ["phy_id", "phy_name", "ifindex", "addr", "type"])
+        phys = []
+        first = True
+
+        # init vars
+        phy_id = ""
+        phy_name = ""
+        ifindex = ""
+        addr = ""
+        _type = ""
+        for line, last_line in flag_last_object(iw_devs.splitlines()):
+            # first phy
+            line = line.strip().lower()
+            if first:
+                if line.startswith("phy#"):
+                    first = False
+                    phy_id = line.split("#")[1]
+                    continue
+            if line.startswith("interface "):
+                phy_name = line.split(" ")[1]
+                continue
+            if line.startswith("ifindex "):
+                ifindex = line.split(" ")[1]
+                continue
+            if line.startswith("addr "):
+                addr = line.split(" ")[1]
+                continue
+            if line.startswith("type "):
+                _type = line.split(" ")[1]
+                continue
+            if line.startswith("wiphy "):
+                phy_id = line.split(" ")[1]
+                continue
+            if line.startswith("phy#"):
+                phys.append(phy(phy_id, phy_name, ifindex, addr, _type))
+                # reset vars
+                phy_id = None
+                phy_name = ""
+                ifindex = None
+                addr = ""
+                _type = ""
+                # new phy
+                phy_id = line.split("#")[1].strip()
+
+            # last phy
+            if last_line:
+                phys.append(phy(phy_id, phy_name, ifindex, addr, _type))
+        return phys
+
+    def get_phy_id(self) -> str:
+        phys = self.build_iw_phy_list()
+        phy_id = ""
+        for phy in phys:
+            if self.name in phy.phy_name:
+                phy_id = phy.phy_id
+        return phy_id
 
     def get_mode(self) -> str:
         """Get the current mode of the interface"""